backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
python-dotenv==1.0.1
python-multipart==0.0.17
httpx==0.27.2
openai==1.55.3
aiofiles==24.1.0
